correcoes

+++ Fase3/src/BussinessLayer/SubCarro/GT.java
@@ -1,24 +1,31 @@
 package BussinessLayer.SubCarro;
 
+import java.util.Random;
+
 public class GT extends Carro{
-    private int taxaDeterioracao;
+    private double taxaDeterioracao;
     
 
-    public GT(double downforce, double fiabilidade, int estadoPneus, String marca, String modelo, int carID, int potenciaICE, int cilindrada, String categoria, TipoPneus tipoPneus, ModoMotor modoMotor,int taxaDeterioracao) {
-        super(downforce, fiabilidade, estadoPneus, marca, modelo, carID, potenciaICE, cilindrada, categoria, tipoPneus, modoMotor);
+    public GT(double downforce, double fiabilidade, int estadoPneus, String marca, String modelo, 
+                int carID, int potenciaICE, int cilindrada, String categoria, TipoPneus tipoPneus, 
+                ModoMotor modoMotor,double taxaDeterioracao) {
+
+        super(downforce, fiabilidade, estadoPneus, marca, modelo, carID, 
+                potenciaICE, cilindrada, categoria, tipoPneus, modoMotor);
+
         this.taxaDeterioracao = taxaDeterioracao;
     }
 
-    public GT(Carro carro, int taxaDeterioracao) {
+    public GT(Carro carro, double taxaDeterioracao) {
         super(carro);
         this.taxaDeterioracao = taxaDeterioracao;
     }
- 
-    public int getTaxaDeterioracao() {
+
+    public double getTaxaDeterioracao() {
         return taxaDeterioracao;
     }
 
-    public void setTaxaDeterioracao(int taxaDeterioracao) {
+    public void setTaxaDeterioracao(double taxaDeterioracao) {
         this.taxaDeterioracao = taxaDeterioracao;
     }
 
@@ -31,9 +38,5 @@ public class GT extends Carro{
 
         this.setTaxaDeterioracao(num);
     }
-    
-
-
-
 
 }
